backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
python-dotenv
langchain-openai
openai
chromadb 
langchain-community 
tiktoken
langchain_chroma
pypdf